+++ mongoose/controllers/holidayController.js
@@ -1,5 +1,4 @@
 const path = require("path");
-const moment = require("moment-timezone");
 const logger = require("../../services/loggerService");
 const holidayService = require("../services/holidayService");
 const mdb = require("../services/mongooseDatabaseService");