+++ mongoose/controllers/twoFAController.js
@@ -2,7 +2,6 @@ const mongoose = require("mongoose");
 const path = require("path");
 const mdb = require("../services/mongooseDatabaseService");
 const logger = require("../../services/loggerService");
-const moment = require("moment-timezone");
 const encryptionService = require("../../services/encryptionService");
 
 function getSafeNext(raw) {